layout routes & Refactoring Route Components
