+++ backend/posts.json
@@ -1,8 +1,8 @@
 {
   "posts": [
     {
-      "body": "This course is hopefully very helpful to you!",
-      "author": "Maximilian",
+      "body": "My default test post!",
+      "author": "Anas Ahmed",
       "id": "post-1"
     }
   ]
