+++ src/main.jsx
@@ -1,14 +1,25 @@
 import React from 'react'
 import ReactDOM from 'react-dom/client'
-import App from './App'
 import './index.css'
 import { createBrowserRouter, RouterProvider } from 'react-router-dom'
-import NewPost from './components/NewPost'
+import NewPost from './routes/NewPost'
+import RootLayout from './routes/RootLayout'
+import Posts from './routes/Posts'
 
 const router = createBrowserRouter([
-  { path: '/', element: <App /> },
-  {path: 'new-post', element: <NewPost />}
-])
+  {
+    path: '/',
+    element: <RootLayout />,
+    children: [
+      {
+        path: '/',
+        element: <Posts />,
+        children: [{ path: 'new-post', element: <NewPost /> }],
+      },
+    ],
+  },
+]);
+
 ReactDOM.createRoot(document.getElementById('root')).render(
   <React.StrictMode>
     <RouterProvider router={router} />
