+++ src/routes/Posts.jsx
@@ -1,0 +1,15 @@
+import { Outlet } from "react-router-dom";
+import PostsList from "../components/PostsList";
+
+function Posts() {
+  return (
+    <>
+      <Outlet />
+      <main>
+        <PostsList />
+      </main>
+    </>
+  )
+}
+
+export default Posts;
