+++ src/routes/RootLayout.jsx
@@ -1,0 +1,14 @@
+import React from 'react'
+import MainHeader from '../components/MainHeader'
+import { Outlet } from 'react-router-dom'
+
+function RootLayout() {
+    return (
+        <>
+            <MainHeader />
+            <Outlet />
+        </>
+    )
+}
+
+export default RootLayout
